chang

+++ apistudy/Models/Entityies/prouduct.cs
@@ -4,13 +4,13 @@
     {
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         //public string Image { get; set; }
         public int CategoryId { get; set; }
         public virtual Category Category { get; set; }
         public virtual ICollection<ProductImage> ProductImages { get; set; }
         public int Quantity { get; set; }
-        public decimal Offer { get; set; }
+        public double Offer { get; set; }
 
 
     }
